+++ CHANGELOG.md
@@ -5,11 +5,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [4.3.0] - 2021-08-09
+### Added
+- `.floor` & `.ceil` as delegates to `.lowerBound` & `upperBound`.
+
+### Fixed
+- `.lowerBound` & `upperBound` now finds the precise bound when multiple ones exist.
+- make param (value) optional on `.insert`.
+
 ## [4.2.2] - 2021-06-20
 
 ### Fixed
 - index.d.ts
-
 
 ## [4.2.1] - 2021-06-20